refactor: move RPC calls to rpcInterface.ts
