+++ lean4-infoview/src/infoview/util.ts
@@ -279,6 +279,17 @@ export function mapRpcError(err : unknown) : Error {
     }
 }
 
+/** Catch handler for RPC methods that just returns undefined if the method is not found.
+ * This is useful for compatibility with versions of Lean that do not yet have the given RPC method.
+*/
+ export function discardMethodNotFound(e: unknown) : undefined {
+  if (isRpcError(e) && (e.code === RpcErrorCode.MethodNotFound)) {
+    return undefined
+  } else {
+      throw mapRpcError(e)
+  }
+}
+
 type Status = 'pending' | 'fulfilled' | 'rejected'
